+++ django/applications/catmaid/control/neuron_annotations.py
@@ -467,36 +467,46 @@ def annotate_entities(request, project_id = None):
     return HttpResponse(json.dumps(result), content_type='text/json')
 
 @requires_user_role(UserRole.Annotate)
-def remove_annotation(request, project_id=None, annotation_id=None):
+def remove_annotations(request, project_id=None):
     """ Removes an annotation from one or more entities.
     """
-    p = get_object_or_404(Project, pk=project_id)
-
+    annotation_ids = [int(v) for k,v in request.POST.iteritems()
+            if k.startswith('annotation_ids[')]
     entity_ids = [int(v) for k,v in request.POST.iteritems()
             if k.startswith('entity_ids[')]
 
-    # Get CICI instance representing the link
-    cici_n_a = ClassInstanceClassInstance.objects.filter(project=p,
-            class_instance_a__id__in=entity_ids,
-            class_instance_b__id=annotation_id)
-    # Make sure the current user has permissions to remove the annotation.
-    missed_cicis = []
-    cicis_to_delete = []
-    for cici in cici_n_a:
-        try:
-            can_edit_or_fail(request.user, cici.id,
-                             'class_instance_class_instance')
-            cicis_to_delete.append(cici)
-        except Exception:
-            # Remember links for which permissions are missing
-            missed_cicis.append(cici)
+    if not annotation_ids:
+        raise ValueError("No annotation IDs provided")
 
-    # Remove link between entity and annotation for all links on which the user
-    # the necessary permissions has.
-    if cicis_to_delete:
-        ClassInstanceClassInstance.objects \
-                .filter(id__in=[cici.id for cici in cicis_to_delete]) \
-                .delete()
+    if not entity_ids:
+        raise ValueError("No entity IDs provided")
+
+    # Remove individual annotations
+    deleted_annotations = []
+    num_left_annotations = {}
+    for annotation_id in annotation_ids:
+        cicis_to_delete, missed_cicis, deleted, num_left = _remove_annotation(
+                request.user, project_id, entity_ids, annotation_id)
+        # Keep track of results
+        num_left_annotations[str(annotation_id)] = num_left
+        for cici in cicis_to_delete:
+            deleted_annotations.append(cici.id)
+
+    return HttpResponse(json.dumps({
+        'deleted_annotations': deleted_annotations,
+        'left_uses': num_left_annotations
+    }), content_type='text/json')
+
+
+@requires_user_role(UserRole.Annotate)
+def remove_annotation(request, project_id=None, annotation_id=None):
+    """ Removes an annotation from one or more entities.
+    """
+    entity_ids = [int(v) for k,v in request.POST.iteritems()
+            if k.startswith('entity_ids[')]
+
+    cicis_to_delete, missed_cicis, deleted, num_left = _remove_annotation(
+            request.user, project_id, entity_ids, annotation_id)
 
     if len(cicis_to_delete) > 1:
         message = "Removed annotation from %s entities." % len(cicis_to_delete)
@@ -509,12 +519,6 @@ def remove_annotation(request, project_id=None, annotation_id=None):
         message += " Couldn't de-annotate %s entities, due to the lack of " \
                 "permissions." % len(missed_cicis)
 
-    # Remove the annotation class instance, regardless of the owner, if there
-    # are no more links to it
-    annotated_with = Relation.objects.get(project_id=project_id,
-            relation_name='annotated_with')
-    deleted, num_left = delete_annotation_if_unused(project_id, annotation_id,
-                                                    annotated_with)
     if deleted:
         message += " Also removed annotation instance, because it isn't used " \
                 "anywhere else."
@@ -526,6 +530,50 @@ def remove_annotation(request, project_id=None, annotation_id=None):
         'deleted_annotation': deleted,
         'left_uses': num_left
     }), content_type='text/json')
+
+def _remove_annotation(user, project_id, entity_ids, annotation_id):
+    """Remove an annotation made by a certain user in a given project on a set
+    of entities (usually neurons and annotations). Returned is a 4-tuple which
+    holds the deleted annotation links, the list of links that couldn't be
+    deleted due to lack of permission, if the annotation itself was removed
+    (because it wasn't used anymore) and how many uses of this annotation are
+    left.
+    """
+    p = get_object_or_404(Project, pk=project_id)
+    relations = dict(Relation.objects.filter(
+        project_id=project_id).values_list('relation_name', 'id'))
+
+    # Get CICI instance representing the link
+    cici_n_a = ClassInstanceClassInstance.objects.filter(project=p,
+            relation_id=relations['annotated_with'],
+            class_instance_a__id__in=entity_ids,
+            class_instance_b__id=annotation_id)
+    # Make sure the current user has permissions to remove the annotation.
+    missed_cicis = []
+    cicis_to_delete = []
+    for cici in cici_n_a:
+        try:
+            can_edit_or_fail(user, cici.id, 'class_instance_class_instance')
+            cicis_to_delete.append(cici)
+        except Exception:
+            # Remember links for which permissions are missing
+            missed_cicis.append(cici)
+
+    # Remove link between entity and annotation for all links on which the user
+    # the necessary permissions has.
+    if cicis_to_delete:
+        ClassInstanceClassInstance.objects \
+                .filter(id__in=[cici.id for cici in cicis_to_delete]) \
+                .delete()
+
+    # Remove the annotation class instance, regardless of the owner, if there
+    # are no more links to it
+    annotated_with = Relation.objects.get(project_id=project_id,
+            relation_name='annotated_with')
+    deleted, num_left = delete_annotation_if_unused(project_id, annotation_id,
+                                                    annotated_with)
+
+    return cicis_to_delete, missed_cicis, deleted, num_left
 
 def create_annotation_query(project_id, param_dict):